--- src/components/Toolbar.css
@@ -1,46 +1,0 @@
-.toolbar{
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    background: #C12584;
-    height: 150px;  
-}
-
-.toolbar__navigation {
-    display: flex;
-    height: 100%;
-    align-items: center;
-    padding: 0 1rem;
-}
-
-.toolbar_logo a {
-    color: white;
-    text-decoration: none;
-    font-size: 1.5rem;
-}
-
-.spacer {
-    flex: 0.5;
-}
-
-.toolbar_navigation-items ul {
-    list-style:none;
-    margin: 0;
-    padding: 0;
-    display: flex;
-}
-
-.toolbar_navigation-items li {
-    padding: 0 0.5rem;
-}
-
-.toolbar_navigation-items a {
-    color: white;
-    text-decoration: none;
-}
-
-.toolbar_navigation-items a:hover,
-.toolbar_navigation-items a:active {
-    color: #00BCD4;
-}
